Try swapping order of gr-leo outputs
Probable bug in rattler-build: when it does the cache build, it seems to
act like it's building for the first output, which in this case is
noarch and therefore activation scripts don't get run when entering the
cache host/build environments. The osx builds don't like this, but linux
manages anyway.

Maybe by changing the order in the recipe, I can get it to build the
non-noarch output first and avoid the bug.