Force skip tests on linux-ppc64le since it is cross-compiled
Using --test=native mostly works for cross-compilation by not running
the tests, but it fails when a noarch package is built because it still
will try to run the tests but does so for the host platform, leading to
failure when not emulating. Something is broken in the logic of
rattler-build, but I can't work out exactly what should change.